Criada a autenticação utilizando Sanctum

+++ routes/api.php
@@ -23,11 +23,17 @@ Route::get('/ping', function(){
     ];
 });
 
-Route::post('/user', [AuthController::class, 'create']);
+Route::get('unauthenticated', function(){
+    return ['error' => 'Usuário não logado!'];
+})->name('login');
 
-Route::post('/todo', [ApiController::class, 'createTodo']);
+Route::post('/user', [AuthController::class, 'create']);
+Route::middleware('auth:sanctum')->get('/auth/logout', [AuthController::class, 'logout']);
+Route::post('/auth', [AuthController::class, 'login']);
+
+Route::middleware('auth:sanctum')->post('/todo', [ApiController::class, 'createTodo']);
 Route::get('/todos', [ApiController::class, 'readlAlltodos']);
 Route::get('/todo/{id}', [ApiController::class, 'readTodo']);
-Route::put('/todo/{id}', [ApiController::class, 'updateTodo']);
-Route::delete('todo/{id}', [ApiController::class, 'deleteTodo']);
+Route::middleware('auth:sanctum')->put('/todo/{id}', [ApiController::class, 'updateTodo']);
+Route::middleware('auth:sanctum')->delete('/todo/{id}', [ApiController::class, 'deleteTodo']);
